+++ bobcoin-consensus/Block.js
@@ -1,17 +1,17 @@
 import { hash, sign, verify } from './cryptoUtils.js';
 
 export class Block {
-    constructor({ type, account, previous, balance, link }) {
+    constructor({ type, account, previous, balance, link, spora = null }) {
         this.type = type;         // 'open', 'send', 'receive'
         this.account = account;   // Public key of the chain owner
         this.previous = previous; // Hash of the previous block (null if 'open')
         this.balance = balance;   // Resulting balance after this block
         this.link = link;         // 'send' -> destination account, 'receive' -> send block hash
+        this.spora = spora;       // SPoRA proof object: { infoHash, challenge, chunkHash }
         this.timestamp = Date.now();
         
         this.hash = this.calculateHash();
         this.signature = null;
-        this.work = null;         // Anti-spam PoW
     }
 
     calculateHash() {
@@ -20,7 +20,8 @@ export class Block {
             this.account +
             (this.previous || '') +
             this.balance.toString() +
-            this.link
+            this.link +
+            (this.spora ? JSON.stringify(this.spora) : '')
         );
     }